add stats syscall

+++ orbit-core/tests/affinity.rs
@@ -186,7 +186,7 @@ fn views<'a>(slots: &'a [AtomicPtr<()>]) -> (HartView<'a>, Vec<HartView<'a>>) {
         .iter()
         .enumerate()
         .map(|(i, slot)| HartView {
-            hart_id: (i + 1) as u32,
+            hart_id: (i + 1),
             current: slot,
         })
         .collect();
